Update boxesitemstyle.md
I made minor copy editing changes. In my opinion, code examples and figures should be labeled with SEO-friendly captions so they are more easily found with Google search--that is how the majority of customers find your help articles. Then in the body copy before the examples or figures, you can add text such as "In **Example 1** you can see x, y and z happen when you set the A and B properties, and you can see the change in **Figure 2**." You're really trying to make the article as easy to find by Google search as possible, so as you label examples and figures, think to yourself, "What would a customer type into search to find this specific example or figure?" and use that in your caption.